feat(install): schematics
