+++ packages/install/src/cli.ts
@@ -1,26 +1,21 @@
 import chalk from 'chalk';
-import resolvePath from 'resolve-package-path';
-import { PackageParser } from '@package/parsers/package.parser';
 import { NpmInstaller } from '@package/installers/npm.installer';
 import { ArgumentParser } from '@package/parsers/argument.parser';
+import { resolvePackageName } from '@package/utils/package.utils';
+import { SchemaInstaller } from '@package/installers/schema.installer';
 
 (async () => {
-  const args = (new ArgumentParser()).parse();
-  const packageName = PackageParser.createName(args.type, args.name);
-
   try {
+    const args = (new ArgumentParser()).parse();
+    const packageName = resolvePackageName(args.type, args.name);
+
     console.info(chalk.blueBright(`Installing package ${packageName}`));
-
     await NpmInstaller.installPackage(packageName, args.path);
-    const packageNameWithoutVersion = PackageParser.stripVersion(packageName);
-    const pathToPackageJson = resolvePath(packageNameWithoutVersion, args.path);
-    const peerDependencies = await (new PackageParser()).parse(pathToPackageJson);
 
-    console.info(chalk.blue('Installing peer dependencies:'));
-    peerDependencies.map(dep => console.info(chalk.blue(`  - ${dep}`)));
-    await NpmInstaller.installPackages(peerDependencies, args.path);
+    console.info(chalk.blue('Running installation schema'));
+    await SchemaInstaller.install(packageName);
 
-    console.info(chalk.bold.green('All packages have been installed.'));
+    console.info(chalk.bold.green(`Package ${packageName} have been installed.`));
   } catch (e) {
     console.error(chalk.bold.red(`${e}`));
   }
