+++ packages/install/src/installers/npm.installer.spec.ts
@@ -19,22 +19,6 @@ jest.mock('child_process', () => ({
       const error = new CustomExecException(1, 'Fail 1');
       cb(error, '', error.message);
     })
-    .mockImplementationOnce((
-      cmd: string,
-      opt: { cmd: string },
-      cb: (err: ExecException, _: string, stderr: string) => void,
-    ) => {
-      const error = new CustomExecException(0, 'All good');
-      cb(error, '', error.message);
-    })
-    .mockImplementationOnce((
-      cmd: string,
-      opt: { cmd: string },
-      cb: (err: ExecException, _: string, stderr: string) => void,
-    ) => {
-      const error = new CustomExecException(2, 'Fail 2');
-      cb(error, '', error.message);
-    }),
 }));
 
 class CustomExecException implements ExecException {
@@ -55,7 +39,7 @@ describe('NPM Installer', () => {
         expect(result).toBe(undefined);
         expect(spy).toHaveBeenNthCalledWith(
           1,
-          'npm install --save-dev @namespace/package --legacy-peer-deps',
+          'npm install --save-dev @namespace/package',
           expect.anything(),
           expect.anything(),
         );
@@ -66,29 +50,6 @@ describe('NPM Installer', () => {
     it('should fail the package installation on error', (done) => {
       NpmInstaller.installPackage('@namespace/package', '/abs/path').catch(result => {
         expect(result).toBe('Fail 1');
-        done();
-      });
-    });
-  });
-
-  describe('installPackages', () => {
-    it('should install multiple packages', (done) => {
-      NpmInstaller.installPackages(['@namespace/package@">= 12.0.0"'], '/abs/path').then(result => {
-        const spy = jest.requireMock('child_process').exec;
-        expect(result).toBe(undefined);
-        expect(spy).toHaveBeenNthCalledWith(
-          3,
-          'npm install --save-dev @namespace/package@">= 12.0.0"',
-          expect.anything(),
-          expect.anything(),
-        );
-        done();
-      });
-    });
-
-    it('should fail the package installation on error', (done) => {
-      NpmInstaller.installPackages(['@namespace/package@">= 12.0.0"'], '/abs/path').catch(result => {
-        expect(result).toBe('Fail 2');
         done();
       });
     });
