+++ packages/install/src/installers/npm.installer.ts
@@ -5,7 +5,7 @@ export class NpmInstaller {
     const saveAsArg = `--save${saveAs === 'dependency' ? '' : '-dev'}`;
     return new Promise(async (res, rej) => {
       exec(
-        `npm install ${saveAsArg} ${packageName} --legacy-peer-deps`,
+        `npm install ${saveAsArg} ${packageName}`,
         { cwd },
         (err, _, stderr) => {
           const errorCode = err?.code ?? 0;
@@ -14,20 +14,6 @@ export class NpmInstaller {
           }
 
           res();
-      });
-    });
-  }
-
-  static installPackages(packages: string[], cwd: string, saveAs: 'devDependency' | 'dependency' = 'devDependency'): Promise<void> {
-    const saveAsArg = `--save${saveAs === 'dependency' ? '' : '-dev'}`;
-    return new Promise(async (res, rej) => {
-      exec(`npm install ${saveAsArg} ${packages.join(' ')}`, { cwd }, (err, _, stderr) => {
-        const errorCode = err?.code ?? 0;
-        if (errorCode > 0) {
-          return rej(stderr);
-        }
-
-        res();
       });
     });
   }
